Added panning base gesture
